shapeai/services/ai-engine/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
mediapipe==0.10.14
anthropic==0.40.0
boto3==1.35.73
psycopg2-binary==2.9.10
requests==2.32.3
pillow==11.0.0
numpy==1.26.4
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
